# apps/api/src/openapi.yaml
openapi: 3.0.3
info:
  title: QiFlow API
  description: |
    The QiFlow payment gateway API. Integrate once, accept Qi everywhere.

    ## Authentication

    QiFlow uses two authentication mechanisms depending on the endpoint group:

    ### JWT Bearer Token (Dashboard endpoints)
    Issued on login / register. Short-lived access token (15 min) + long-lived refresh token (7 days).
    ```
    Authorization: Bearer <accessToken>
    ```

    ### API Key (Programmatic / v1 endpoints)
    Generated from the dashboard or via `POST /merchants/me/api-keys`.
    ```
    X-API-Key: qf_test_xxxxxxxxxxxxxxxx
    ```

    ## Rate Limiting
    Rate-limit headers are returned on every response:
    | Header | Description |
    |---|---|
    | `RateLimit-Limit` | Max requests per window |
    | `RateLimit-Remaining` | Requests left in current window |
    | `RateLimit-Reset` | Seconds until window resets |

    Auth endpoints have stricter limits (5 req/min for login, 3 req/min for register).
    All other endpoints share a global limit of 500 req/15 min.

    ## Error Format
    All errors follow a consistent shape:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable message"
      }
    }
    ```
  version: 1.0.0
  contact:
    name: QiFlow
    url: https://qiflow.xyz
servers:
  - url: https://api.qiflow.xyz
    description: Production
  - url: http://localhost:3001
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from `/auth/login` or `/auth/register`
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key obtained from the merchant dashboard or `POST /merchants/me/api-keys`

  # ── Reusable Schemas ─────────────────────────────────────────────────────────
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: INVALID_CREDENTIALS
            message:
              type: string
              example: Invalid email or password
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        details:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email: ["Invalid email address"]
            password: ["Password must be at least 8 characters"]

    # ── Tokens ───────────────────────────────────────────────────────────────────
...
    Webhook:
      type: object
      required: [id, url, events, isActive, createdAt]
      properties:
        id:
          type: string
          format: uuid
          example: 9b2d5e8f-1234-4abc-8765-abcdef012345
        url:
          type: string
          format: uri
          example: https://myshop.com/webhooks/qiflow
        events:
          type: array
          items:
            type: string
            enum: [payment.created, payment.pending, payment.completed, payment.failed, payment.expired]
          example: [payment.completed, payment.failed]
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2026-01-01T00:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-01-01T00:00:00.000Z"

    # ── Stats ────────────────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Missing or invalid Authorization header
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
paths:
...
  /v1/webhooks:
    post:
      tags: [Webhooks]
      summary: Create a webhook
      description: |
        Registers a new webhook endpoint. QiFlow will send signed HTTP POST
        requests to this URL when the specified events occur.

        **Authentication:** JWT Bearer token
      operationId: createWebhook
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url:
                  type: string
                  format: uri
                  example: https://myshop.com/webhooks/qiflow
                events:
                  type: array
                  minItems: 1
                  items:
                    type: string
                    enum: [payment.created, payment.pending, payment.completed, payment.failed, payment.expired]
                  example: [payment.completed, payment.failed]
            example:
              url: https://myshop.com/webhooks/qiflow
              events: [payment.completed, payment.failed]
      responses:
        '201':
          description: Webhook created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Webhook'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags: [Webhooks]
      summary: List webhooks
      description: |
        Returns all webhook endpoints for the authenticated merchant.

        **Authentication:** JWT Bearer token
      operationId: listWebhooks
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of webhooks
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Webhook'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/webhooks/{id}:
    put:
      tags: [Webhooks]
      summary: Update a webhook
      description: |
        Updates the URL, events, or active status of a webhook.

        **Authentication:** JWT Bearer token
      operationId: updateWebhook
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: 9b2d5e8f-1234-4abc-8765-abcdef012345
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                  example: https://myshop.com/webhooks/qiflow-v2
                events:
                  type: array
                  items:
                    type: string
                    enum: [payment.created, payment.pending, payment.completed, payment.failed, payment.expired]
                isActive:
                  type: boolean
                  example: true
      responses:
        '200':
          description: Webhook updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Webhook'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Webhooks]
      summary: Delete a webhook
      description: |
        Permanently removes a webhook endpoint.

        **Authentication:** JWT Bearer token
      operationId: deleteWebhook
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: 9b2d5e8f-1234-4abc-8765-abcdef012345
      responses:
        '200':
          description: Webhook deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      message:
                        type: string
                        example: Webhook deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Merchants ─────────────────────────────────────────────────────────────────
...